pyproject.toml
[project]
name = "boosta-task"
version = "0.1.0"
description = "Voonix API utilization with GCP tools project"
authors = [
    {name = "vvk",email = "koval.vladislavv@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "polars (>=1.30.0,<2.0.0)",
    "google-cloud-storage (>=3.1.0,<4.0.0)",
    "google-cloud-bigquery (>=3.34.0,<4.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "ruff (>=0.11.13,<0.12.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
